[MODIFY] 레포지토리 이름 변경
