--- calculator/src/main/java/repository/MemoryRepository.java
@@ -1,13 +1,0 @@
-package repository;
-
-import entity.Expression;
-
-import java.util.List;
-import java.util.Optional;
-
-public interface MemoryRepository {
-
-    Expression save(String input, double result);
-    List<Expression> findAll();
-
-}
